Add option to enable/disable offensive words patrol.

+++ src/main/java/geniusBot/offensiveWordsPatrol/OffensiveWordsPatrol.java
@@ -19,30 +19,79 @@
 
 package geniusBot.offensiveWordsPatrol;
 
+import geniusBot.mySQLutils.MySQLConnection;
 import net.dv8tion.jda.api.entities.Message;
 import net.dv8tion.jda.api.events.message.guild.GuildMessageReceivedEvent;
 import net.dv8tion.jda.api.hooks.ListenerAdapter;
 
+import javax.sql.rowset.RowSetWarning;
+import java.sql.ResultSet;
+import java.sql.SQLException;
+import java.sql.Connection;
+import java.sql.Statement;
+
 public class OffensiveWordsPatrol extends ListenerAdapter
 {
    public static final String[] BAD_WORDS = {"fuck", "bitch", "shit"};
+   public static final String SQL_OFFENSIVE_WORDS_PATROL_GUILD_DATA_TALE =
+           "offensive_words_patrol_guild_data";
 
    @Override
    public void onGuildMessageReceived(GuildMessageReceivedEvent event)
    {
       Message msg = event.getMessage();
+      long guildId = msg.getGuild().getIdLong();
+      boolean[] enabledData = checkGuildOffensiveWordsPatrolEnabled(guildId);
+      // return immediately of offensive words patrol isn't enabled
+      if (!enabledData[0] || (msg.getTextChannel().isNSFW() && !enabledData[1]))
+         return;
       String msgData = msg.getContentRaw();
       msgData = msgData.toLowerCase();
       for(String str : BAD_WORDS)
-      { 
+      {
          if(msgData.indexOf(str) != -1)
          {
             msg.delete().queue();
-            event.getAuthor().openPrivateChannel().flatMap(channel -> 
-            channel.sendMessage("<@" + event.getAuthor().getId() + "> " + 
+            event.getAuthor().openPrivateChannel().flatMap(channel ->
+            channel.sendMessage("<@" + event.getAuthor().getId() + "> " +
                   "Please don't be rude in this server! :angry:")).queue();
             break;
          }
+      }
+   }
+
+   // Return boolean array: pos 0 identifies whether the patrol is enabled, and
+   // pos 1 identifies if it is enabled on nsfw channels;
+   public static boolean[] checkGuildOffensiveWordsPatrolEnabled(long guildId)
+   {
+      try
+      {
+         Connection con = MySQLConnection.getConnection();
+         Statement statement = con.createStatement();
+         ResultSet resultSet = statement.executeQuery("SELECT watchNSFW FROM "
+                 + SQL_OFFENSIVE_WORDS_PATROL_GUILD_DATA_TALE + " WHERE guildId = \""
+                 + guildId + "\";");
+
+         boolean result[] = new boolean[2];
+         if (resultSet.next())
+         {
+            result[0] = true;
+            boolean watchNSFW = resultSet.getBoolean(1);
+            result[1] = watchNSFW;
+         }
+         else
+         {
+            result[0] = false;
+            result[1] = false;
+         }
+
+         con.close();
+         return result;
+      }
+      catch (SQLException e)
+      {
+         e.printStackTrace();
+         return new boolean[]{false, false};
       }
    }
 }
